put authorization restriction on API, must be logged in to view & post, only author of post can delete or put it, NOTE : doesn't handle friends/visibility/privacy authorization yet

+++ api/views.py
@@ -23,24 +23,34 @@ class PostList(generics.GenericAPIView):
 	queryset = Post.objects.all()
 
 	def get(self, request, format=None):
-		posts = Post.objects.filter(visibility='PUBLIC').order_by('-published')
-		page = self.paginate_queryset(posts)
-		if page is not None:
-			serializer = PostSerializer(page, many=True)
-			return self.get_paginated_response({"data":serializer.data, "query": "posts"})
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
+			posts = Post.objects.filter(visibility='PUBLIC').order_by('-published')
+			page = self.paginate_queryset(posts)
+			if page is not None:
+				serializer = PostSerializer(page, many=True)
+				return self.get_paginated_response({"data":serializer.data, "query": "posts"})
+			#else:
 
-		# else 
+		else:
+			return Response(serializer.errors, status=HTTP_401_UNAUTHORIZED)
 
 	def post(self, request, format=None):
-		serializer = PostSerializer(data=request.data)
-		if serializer.is_valid():
-			print "DEBUG : API - views.py - PostList"
-			serializer.validated_data["author"] = request.user
-			serializer.validated_data["published"] = timezone.now()
-			serializer.save()
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
+			serializer = PostSerializer(data=request.data)
+			if serializer.is_valid():
+				print "DEBUG : API - views.py - PostList"
+				serializer.validated_data["author"] = request.user
+				serializer.validated_data["published"] = timezone.now()
+				serializer.save()
+				return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-			return Response(serializer.data, status=status.HTTP_201_CREATED)
-		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+			else:
+				Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+		else:
+			return Response(serializer.errors, status=HTTP_401_UNAUTHORIZED)
 
 	def perform_create(self, serializer):
 		serializer.save(author=self.request.user)
@@ -62,26 +72,53 @@ class PostDetail(generics.GenericAPIView):
 
 
 	def get(self, request, pk, format=None):
-		post = self.get_object(pk)
-		serializer = PostSerializer(post)
-		return Response(serializer.data)
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
 
+			# --- TODO : Only authorize users to read/get this post if visibility/privacy settings allow it
+			post = self.get_object(pk)
+			serializer = PostSerializer(post)
+			return Response(serializer.data)
+
+		else:
+			return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 	def put(self, request, pk, format=None):
-		post = self.get_object(pk)
-		serializer = PostSerializer(post, data=request.data)
-		if serializer.is_valid():
-			serializer.save()
-			return Response(serializer.data)
-		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
+			post = self.get_object(pk)
+
+			# only allow author of the post to modify it
+			if request.user == post.author:
+				serializer = PostSerializer(post, data=request.data)
+				if serializer.is_valid():
+					serializer.save()
+					return Response(serializer.data)
+				return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+			# if logged in user is not author of the post
+			else:
+				return Response(status=status.HTTP_403_FORBIDDEN)
+
+		else:
+			return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 	def delete(self, request, pk, format=None):
-		post = self.get_object(pk)
-		post.delete()
-		return Response(status=status.HTTP_204_NO_CONTENT)
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
+			post = self.get_object(pk)
+			# only allow author of the post to modify it
+			if request.user == post.author:
+				post.delete()
+				return Response(status=status.HTTP_204_NO_CONTENT)
 
+			# if logged in user is not author of the post
+			else:
+				return Response(status=status.HTTP_403_FORBIDDEN)
 
+		else:
+			return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 '''
@@ -93,24 +130,39 @@ class CommentList(generics.GenericAPIView):
 	queryset = Comment.objects.all()
 
 	def get(self, request, post_pk, format=None):
-		comments = Comment.objects.filter(post=post_pk).order_by('-published')
-		page = self.paginate_queryset(comments)
-		if page is not None:
-			serializer = CommentSerializer(page, many=True)
-			return self.get_paginated_response({"data":serializer.data, "query": "comments"})
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
+			
+			# --- TODO : Only authorize users to read/get this post if visibility/privacy settings allow it
+			comments = Comment.objects.filter(post=post_pk).order_by('-published')
+			page = self.paginate_queryset(comments)
+			if page is not None:
+				serializer = CommentSerializer(page, many=True)
+				return self.get_paginated_response({"data":serializer.data, "query": "comments"})
+			# else 
 
-		# else 
+		else:
+			return Response(status=status.HTTP_401_UNAUTHORIZED)
+
 
 	def post(self, request, post_pk, format=None):
-		serializer = CommentSerializer(data=request.data)
-		if serializer.is_valid():
-			print "DEBUG : API - views.py - CommentList"
-			serializer.validated_data["author"] = request.user
-			serializer.validated_data["published"] = timezone.now()
-			serializer.validated_data["post"] = Post.objects.get(pk=post_pk)
-			serializer.save()
-			return Response(serializer.data, status=status.HTTP_201_CREATED)
-		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+		# ensure user is authenticated
+		if (request.user.is_authenticated()):
+
+			# -- TODO : Only authorize user who can view the corresponding post to comment
+			serializer = CommentSerializer(data=request.data)
+			if serializer.is_valid():
+				print "DEBUG : API - views.py - CommentList"
+				serializer.validated_data["author"] = request.user
+				serializer.validated_data["published"] = timezone.now()
+				serializer.validated_data["post"] = Post.objects.get(pk=post_pk)
+				serializer.save()
+				return Response(serializer.data, status=status.HTTP_201_CREATED)
+			else:
+				return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+		else:
+			return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 	def perform_create(self, serializer):
 		serializer.save(author=self.request.user)
